+++ src/main/java/pope/two_one_three_seven/model/Point.java
@@ -2,23 +2,23 @@ package pope.two_one_three_seven.model;
 
 public class Point {
 
-    int x;
-    int y;
+    double x;
+    double y;
     int ID;
     boolean isOnCircle;
 
-    public Point(int x, int y, int ID, boolean isOnCircle){
+    public Point(double x, double y, int ID, boolean isOnCircle){
         this.x = x;
         this.y = y;
         this.ID = ID;
         this.isOnCircle = isOnCircle;
     }
 
-    public int getX(){
+    public double getX(){
         return this.x;
     }
 
-    public int getY(){
+    public double getY(){
         return this.y;
     }
